+++ workspaces/npm/plugins/npm/src/alpha.tsx
@@ -15,7 +15,6 @@
  */
 import {
   ApiBlueprint,
-  createApiFactory,
   createFrontendPlugin,
   discoveryApiRef,
   fetchApiRef,
@@ -39,8 +38,8 @@ export { isNpmAvailable } from '@backstage-community/plugin-npm-common';
  */
 export const npmBackendApi = ApiBlueprint.make({
   name: 'npmBackendApi',
-  params: {
-    factory: createApiFactory({
+  params: defineParams =>
+    defineParams({
       api: NpmBackendApiRef,
       deps: {
         discoveryApi: discoveryApiRef,
@@ -52,7 +51,6 @@ export const npmBackendApi = ApiBlueprint.make({
           fetchApi,
         }),
     }),
-  },
 });
 
 /**
@@ -100,8 +98,8 @@ export const entityNpmReleaseOverviewCard: any = EntityCardBlueprint.make({
 export const entityNpmReleaseTableCard: any = EntityContentBlueprint.make({
   name: 'EntityNpmReleaseTableCard',
   params: {
-    defaultPath: 'npm-releases',
-    defaultTitle: 'Npm Releases',
+    path: 'npm-releases',
+    title: 'Npm Releases',
     filter: isNpmAvailable,
     loader: () =>
       import('./components/EntityNpmReleaseTableCard').then(m => (