+++ workspaces/bookmarks/plugins/bookmarks/CHANGELOG.md
@@ -1,7 +1,7 @@
 # @backstage-community/plugin-bookmarks
 
-## 0.2.0
+## 0.1.0
 
 ### Minor Changes
 
-- 80b21b2: Initial release
+- b01eb10: Initial release