+++ workspaces/tech-radar/plugins/tech-radar-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-tech-radar-common
 
+## 1.9.1
+
+### Patch Changes
+
+- 99e80be: Backstage version bump to v1.42.4
+
 ## 1.9.0
 
 ### Minor Changes